+++ CHANGES.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Removed explicit error handler for uncaught exceptions, improving security by preventing sensitive information from being exposed in error responses. ([#966](https://github.com/stac-utils/stac-fastapi/pull/966))
+- Removed deprecated Fastapi's orjson custom response (`ORJSONResponse`) in favor of the default `JSONResponse`
 
 ## [6.5.0] - 2026-08-03