+++ CHANGES.md
@@ -4,7 +4,14 @@
 
 ### Changed
 
+- Changed output from `string` to `dict` object (`BulkTransaction`) for bulk transaction `bulk_item_insert` methods. 
+
+## [6.6.0] - 2026-08-31
+
+### Changed
+
 - Removed explicit error handler for uncaught exceptions, improving security by preventing sensitive information from being exposed in error responses. ([#966](https://github.com/stac-utils/stac-fastapi/pull/966))
+- Removed deprecated Fastapi's orjson custom response (`ORJSONResponse`) in favor of the default `JSONResponse`
 
 ## [6.5.0] - 2026-08-03
 
@@ -777,7 +784,8 @@ Full changelog: https://stac-utils.github.io/stac-fastapi/migrations/v3.0.0/#cha
 
 * First PyPi release!
 
-[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/6.5.0..main>
+[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/6.6.0..main>
+[6.6.0]: <https://github.com/stac-utils/stac-fastapi/compare/6.5.0..6.6.0>
 [6.5.0]: <https://github.com/stac-utils/stac-fastapi/compare/6.4.1..6.5.0>
 [6.4.1]: <https://github.com/stac-utils/stac-fastapi/compare/6.4.0..6.4.1>
 [6.4.0]: <https://github.com/stac-utils/stac-fastapi/compare/6.3.2..6.4.0>